% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{This function will convert between currencies.
authors Carlos Couce, Graham Place}
\usage{
convert(spending_frame, out_curr = "USD")
}
\arguments{
\item{spending_frame}{- data frame passed through input() by user before convert() is called}

\item{out_curr}{- all the currencies in the data frame will
be converted to this output currency}
}
\value{
spending_frame - generate a column at the far right of the data frame
              with the desired output_currency and a running total
}
\description{
This function will convert between currencies.
authors Carlos Couce, Graham Place
}

