% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\docType{class}
\name{Currency-class}
\alias{Currency-class}
\title{An S4 class to represent currency values.}
\description{
Currency is a subclass of "numeric", and is intended to perform currency
conversions, store original unconverted values and the original currency
unit, and provide formatted output in the form of strings like "$33 thousand"
instead of "33000".
}
\section{Slots}{

\describe{
\item{\code{type}}{The original currency unit before conversion.}

\item{\code{original}}{The original currency value before conversion.}
}}

