% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{error_checking}
\alias{error_checking}
\title{This helper function is doing all the necessary error checking to ensure that the data frame can be correctly processed.}
\usage{
error_checking(spending_frame, out_curr)
}
\arguments{
\item{spending_frame}{- data frame passed through input() by user before convert() is called}

\item{out_curr}{- all the currencies in the data frame will
be converted to this output currency}
}
\value{
spending_frame - generate a column at the far right of the data frame
              with the desired output_currency and a running total
}
\description{
authors Carlos Couce, Graham Place
}

