% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{format.Currency}
\alias{format.Currency}
\title{Do fancy printing of currency values.}
\usage{
\method{format}{Currency}(x, original = FALSE, ...)
}
\arguments{
\item{x}{An currency object.}

\item{original}{A boolean indicating whether to print the original,
unconverted value.}
}
\value{
A formatted string
}
\description{
This is currently a proof-of-principle example.
}
