% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Read in spender's spreadsheet and convert to usable data frame
authors Claire Adair, Graham Place}
\usage{
input(csv)
}
\arguments{
\item{csv}{A csv file with expenses information (including currency).}
}
\value{
A data frame with cleaned up and properly identified information.
}
\description{
Read in spender's spreadsheet and convert to usable data frame
authors Claire Adair, Graham Place
}

