% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_bar.R
\name{spending_bar}
\alias{spending_bar}
\title{Display bar chart broken down by spending category
author Graham Place}
\usage{
spending_bar(spending_frame, outputCurr = "USD")
}
\arguments{
\item{spending_frame}{A data frame with expenses information (including currency)}

\item{outputCurr}{desired output currency to display totals in}
}
\description{
Display bar chart broken down by spending category
author Graham Place
}

