% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_by_country.R
\name{spending_by_country}
\alias{spending_by_country}
\title{Display bar chart broken down by spending per country in desired output currency
author Graham Place}
\usage{
spending_by_country(spending_frame, outputCurr = "USD")
}
\arguments{
\item{spending_frame}{A data frame with expenses information (including currency)}

\item{outputCurr}{desired output currency to display totals in}
}
\description{
Display bar chart broken down by spending per country in desired output currency
author Graham Place
}

