% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_by_day.R
\name{spending_by_day}
\alias{spending_by_day}
\title{Display bar chart broken down by day of the week
author Graham Place}
\usage{
spending_by_day(spending_frame, outputCurr = "USD")
}
\arguments{
\item{spending_frame}{A data frame with expenses information (including currency)}

\item{outputCurr}{desired output currency to display totals in}
}
\description{
Display bar chart broken down by day of the week
author Graham Place
}

