% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_line_total.R
\name{spending_line_total}
\alias{spending_line_total}
\title{Display bar chart broken down by spending category
author Graham Place}
\usage{
spending_line_total(spending_frame, outputCurr = "USD")
}
\arguments{
\item{spending_frame}{A data frame with expenses information (including currency)}

\item{outputCurr}{desired output currency to display totals in}
}
\description{
Display bar chart broken down by spending category
author Graham Place
}

