% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Read in spreadsheet and added value, convert and add to data frame
authors Claire Adair, Graham Place}
\usage{
update(csv, currency, value, category, date)
}
\arguments{
\item{csv}{A data frame already passed through input() function.}

\item{currency}{String containing currency code of the added value.}

\item{value}{Integer of expense in original purchase's currency}

\item{category}{String containing the category of expense.}

\item{date}{String containing date of expense in YYYY-MM-DD format.}
}
\value{
A data frame with user inputted data appended at bottom
}
\description{
Read in spreadsheet and added value, convert and add to data frame
authors Claire Adair, Graham Place
}

